python/sk_workflow_example/requirements.txt
requests
python-dotenv
semantic-kernel
edgartools
yfinance==0.2.40